Syntax fixes

+++ ssgan.py
@@ -255,8 +255,8 @@ fixed_noise = torch.FloatTensor(batch_size, nz, 1, 1).normal_(0, 1)
 fixed_noise = _to_var(fixed_noise)
 
 
-d_gan_labels_real = torch.LongTensor(batch_size)
-d_gan_labels_fake = torch.LongTensor(batch_size)
+d_gan_labels_real = torch.FloatTensor(batch_size)
+d_gan_labels_fake = torch.FloatTensor(batch_size)
 
 real_label = 1
 fake_label = 0
@@ -281,7 +281,7 @@ for epoch in range(num_epochs):
 
 
         netD.zero_grad()
-        d_gan_labels_real = d_gan_labels_real.resize_as_(svhn_labels.data.cpu().float()).float().uniform_(0, 0.3)
+        d_gan_labels_real = d_gan_labels_real.resize_as_(svhn_labels.data.cpu().float()).uniform_(0, 0.3)
         d_gan_labels_real_var = _to_var(d_gan_labels_real).float()
         output, _, gan_logits_real, d_sample_features = netD(svhn_data)
 
@@ -303,7 +303,7 @@ for epoch in range(num_epochs):
         noise_var = _to_var(noise)
         fake = netG(noise_var)
 
-        d_gan_labels_fake = d_gan_labels_fake.resize_as_(svhn_labels.data.cpu().float()).float().uniform_(0.9, 1.2)
+        d_gan_labels_fake = d_gan_labels_fake.resize_as_(svhn_labels.data.cpu().float()).uniform_(0.9, 1.2)
         d_gan_labels_fake_var = _to_var(d_gan_labels_fake).float()
         _, _, gan_logits_fake, _ = netD(fake.detach())
